Default Changelist

+++ src/main/java/com/erp/demo/repository/HiringInfoRepository.java
@@ -8,7 +8,7 @@ import java.sql.Date;
 import java.util.List;
 
 @Repository
-public interface HiringInfoRepository extends CrudRepository<HiringInfo, Long> {
+public interface HiringInfoRepository extends CrudRepository<HiringInfo, Integer> {
 
     List<HiringInfo> getAllByTrainingDateAndDocumentsSubmittedAndJoiningDate(Date trainingDate, char documentsSubmitted, Date joiningDate);
     HiringInfo getByHiringId(int hiringId);
